+++ util/ioutils.h
@@ -91,7 +91,7 @@ asprintf_safe(char** strp, const char* format, ...);
 
 int run_command_get_outputs(const char* cmd, sl** outlines, sl** errlines);
 
-void get_mmap_size(int start, int size, off_t* mapstart, size_t* mapsize, int* pgap);
+void get_mmap_size(size_t start, size_t size, off_t* mapstart, size_t* mapsize, int* pgap);
 
 char* resolve_path(const char* filename, const char* basedir);
